ADD: user menu for mobileNavigationBar
