+++ src/shared/ui/Navigation/config/navigationConfig.tsx
