+++ src/shared/ui/Navigation/index.tsx
@@ -1,0 +1,1 @@
+export { Navigation } from "./ui/Navigation";
