+++ src/shared/ui/UserMenu/index.tsx
@@ -1,0 +1,1 @@
+export { UserMenu } from "./ui/UserMenu";
